// pix-front2/src/app/seg/model/parametro.ts
export class Parametro {
  qtdDiaExpSen: number;
  tipCarSen: string;
  tamMinSen: number;
  tamMaxSen: number;
  numCacIguCcu: number;
  qtdMaxAltSenDia: number;
  qtdMinSenCcu: number;
  qtdMinCacLet: number;
  qtdMinLetMau: number;
  qtdMinLetMiu: number;
  qtdMinCacEsp: number;
  qtdMinCacNum: number;
  qtdDiaSemAceSis: number;
  qtdTtvIvlPem: number;
  idEntParInd: number;
  flgPssParInd: string;
  flgUtzDioSenFra: string;
}